chore: renaming test modules
