+++ tests/applications/broken_in_runtime.py
@@ -1,8 +1,12 @@
 from aidial_sdk.chat_completion import ChatCompletion, Request, Response
-from tests.applications.broken_application import raise_exception
+from tests.applications.broken_immediately import raise_exception
 
 
 class RuntimeBrokenApplication(ChatCompletion):
+    """
+    Application which breaks after producing some output.
+    """
+
     async def chat_completion(
         self, request: Request, response: Response
     ) -> None:
