+++ tests/applications/single_choice.py
