+++ tests/test_max_prompt_tokens.py
@@ -3,7 +3,7 @@ from unittest.mock import Mock
 from starlette.testclient import TestClient
 
 from aidial_sdk import DIALApp
-from tests.applications.single_choice_application import SingleChoiceApplication
+from tests.applications.single_choice import SingleChoiceApplication
 
 
 def test_max_prompt_tokens_is_set():
